# swagger.yaml
openapi: 3.0.3
info:
  title: API Venta de Vehículos
  description: |
    API REST desarrollada con **NestJS** y **Sequelize** para gestionar la venta de vehículos.
    
    Permite administrar:
    - 👤 **Clientes**: registro y consulta
    - 🚗 **Vehículos**: CRUD completo
    - 📋 **Solicitudes de crédito**: creación y aprobación
    - 💳 **Cuotas**: consulta y pago
    
    **Base de datos**: SQL Server (`Venta_VehiculoDB`)
  version: 1.0.0
  contact:
    name: Alvin Dalvise & Camila Chanquispe
    email: alvindalviseortiz2310@gmail.com
servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
...
paths:
...
  /solicitudes:
    post:
      tags: [Solicitudes]
      summary: Crear una solicitud de crédito
      operationId: createSolicitud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSolicitudDto'
      responses:
        '201':
          description: Solicitud creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
        '400':
          description: Datos inválidos

    get:
      tags: [Solicitudes]
      summary: Obtener todas las solicitudes
      operationId: getSolicitudes
      responses:
        '200':
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Solicitud'
...
  /cuotas/solicitud/{id}:
    get:
      tags: [Cuotas]
      summary: Obtener cuotas por solicitud
      operationId: getCuotasBySolicitud
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la solicitud
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Cuotas de la solicitud indicada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cuota'
        '404':
          description: Solicitud no encontrada
...
components:
...
  schemas:
...
    Solicitud:
      type: object
      description: Solicitud de crédito para la compra de un vehículo
      properties:
        Id_Solicitud:
          type: integer
          readOnly: true
          example: 1
        Id_Cliente:
          type: integer
          example: 2
        Id_Vehiculo:
          type: integer
          example: 3
        plazoMeses:
          type: integer
          example: 36
        estado:
          type: string
          enum: [PENDIENTE, APROBADA, RECHAZADA]
          example: PENDIENTE
        fechaSolicitud:
          type: string
          format: date
          readOnly: true
          example: '2024-03-16'
    Cuota:
      type: object
      description: Cuota mensual generada para una solicitud aprobada
      properties:
        Id_Cuota:
          type: integer
          readOnly: true
          example: 1
        Id_Solicitud:
          type: integer
          example: 1
        numeroCuota:
          type: integer
          example: 1
        montoCuota:
          type: number
          format: decimal
          example: 347222.22
        fechaVencimiento:
          type: string
          format: date
          example: '2024-04-16'
        pagada:
          type: boolean
          example: false

    # ── DTOs ──────────────────────────────────
...
    CreateSolicitudDto:
      type: object
      required: [Id_Cliente, Id_Vehiculo, plazoMeses]
      properties:
        Id_Cliente:
          type: integer
          example: 1
        Id_Vehiculo:
          type: integer
          example: 2
        plazoMeses:
          type: integer
          minimum: 6
          maximum: 72
          example: 36
...